+++ bundles/org.corpus_tools.hexatomic.core/src/main/java/org/corpus_tools/hexatomic/core/ApplicationLifecycle.java
@@ -25,14 +25,28 @@ import ch.qos.logback.classic.joran.JoranConfigurator;
 import ch.qos.logback.core.joran.spi.JoranException;
 import java.io.File;
 import java.net.URL;
+import javax.inject.Inject;
+import org.corpus_tools.hexatomic.core.errors.ErrorService;
+import org.corpus_tools.hexatomic.core.update.AppStartupCompleteEventHandler;
 import org.eclipse.core.runtime.FileLocator;
+import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.Path;
 import org.eclipse.core.runtime.Platform;
+import org.eclipse.core.runtime.preferences.ConfigurationScope;
+import org.eclipse.core.runtime.preferences.IEclipsePreferences;
+import org.eclipse.e4.core.contexts.IEclipseContext;
+import org.eclipse.e4.core.services.events.IEventBroker;
+import org.eclipse.e4.ui.di.UISynchronize;
+import org.eclipse.e4.ui.workbench.UIEvents;
+import org.eclipse.e4.ui.workbench.lifecycle.PostContextCreate;
 import org.eclipse.e4.ui.workbench.lifecycle.ProcessAdditions;
+import org.eclipse.equinox.p2.core.IProvisioningAgent;
 import org.eclipse.osgi.service.datalocation.Location;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.FrameworkUtil;
+import org.osgi.service.prefs.BackingStoreException;
 import org.slf4j.LoggerFactory;
+
 
 /**
  * Class that hooks into the application lifecycle and configures logging.
@@ -48,7 +62,11 @@ public class ApplicationLifecycle {
 
   private static final org.slf4j.Logger log =
       org.slf4j.LoggerFactory.getLogger(ApplicationLifecycle.class);
-
+  private static final IEclipsePreferences prefs =
+      ConfigurationScope.INSTANCE.getNode("org.corpus_tools.hexatomic.core");
+  @Inject
+  ErrorService errorService;
+  
   /**
    * Called when the model is loaded and initializes the logging.
    */
@@ -79,5 +97,39 @@ public class ApplicationLifecycle {
       log.error("Could not configure logging", ex);
     }
   }
+  
+  @PostContextCreate
+  void postContextcreate(final IProvisioningAgent agent, 
+      UISynchronize sync,
+      IProgressMonitor monitor,
+      IEventBroker eventBroker,
+      IEclipseContext context) {
+    //check if preferences are set to autoupdate and if app was recently updated
+    boolean justUpdated = prefs.getBoolean("justUpdated", false);
+    boolean autoUpdateEnabled = prefs.getBoolean("autoUpdate", true);
+    if (!justUpdated && autoUpdateEnabled) {
+      //add listener to perform updates as soon as workbench is created
+      eventBroker.subscribe(UIEvents.UILifeCycle.APP_STARTUP_COMPLETE, 
+          new AppStartupCompleteEventHandler(eventBroker, 
+            context,
+            agent,
+            sync,
+            monitor,
+              null, eventBroker));
+    } else if (justUpdated) {
+      prefs.putBoolean("justUpdated", false);
+      try {
+        prefs.flush();
+      } catch (BackingStoreException ex) {
+        errorService.handleException("Couldn't update preferences", ex, ApplicationLifecycle.class);
+      }
+    }   
+    
+  }
+  
+
+  
+  
 
 }
+