+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add keyboard shortcut to create spans in Grid Editor (#243)
+- Add "update functionality" in the "Help Menu"
+- Add "Preferences" dialog to allow auto-updates at startup in the "Help Menu"
 - Add overview tables for grid editor keyboard shortcuts to user documentation (#361)
 - Add support for adding new annotation columns via context menu and keyboard shortcuts to Grid Editor (#358)
 - Add command to change the covered text of a token in the Graph Editor (#373)